fix: string extraction from json
